添加权限和性别集合类,完成用户查询功能

+++ src/main/java/com/platform/pojo/User.java
@@ -36,7 +36,7 @@ public class User implements Serializable {
 
     private String picture;
 
-    private Long gender;
+    private Integer gender;
 
     @TableField(fill = FieldFill.INSERT)
     private Date createTime;
